plumb stats through
