+++ rust/core/src/serde/scheduler/from_proto.rs
@@ -17,7 +17,9 @@ use std::{collections::HashMap, convert::TryInto};
 use crate::error::BallistaError;
 use crate::serde::protobuf;
 use crate::serde::protobuf::action::ActionType;
-use crate::serde::scheduler::{Action, ExecutePartition, PartitionId, PartitionLocation};
+use crate::serde::scheduler::{
+    Action, ExecutePartition, PartitionId, PartitionLocation, PartitionStats,
+};
 
 use datafusion::logical_plan::LogicalPlan;
 use uuid::Uuid;
@@ -66,6 +68,12 @@ impl TryInto<PartitionId> for protobuf::PartitionId {
     }
 }
 
+impl Into<PartitionStats> for protobuf::PartitionStats {
+    fn into(self) -> PartitionStats {
+        PartitionStats::new(self.num_rows, self.num_batches, self.num_bytes)
+    }
+}
+
 impl TryInto<PartitionLocation> for protobuf::PartitionLocation {
     type Error = BallistaError;
 
@@ -84,6 +92,14 @@ impl TryInto<PartitionLocation> for protobuf::PartitionLocation {
                 .ok_or_else(|| {
                     BallistaError::General(
                         "executor_meta in PartitionLocation is missing".to_owned(),
+                    )
+                })?
+                .into(),
+            partition_stats: self
+                .partition_stats
+                .ok_or_else(|| {
+                    BallistaError::General(
+                        "partition_stats in PartitionLocation is missing".to_owned(),
                     )
                 })?
                 .into(),
