+++ rust/core/src/serde/scheduler/mod.rs
@@ -14,13 +14,15 @@
 
 use std::{collections::HashMap, sync::Arc};
 
-use arrow::datatypes::{Schema, SchemaRef};
+use arrow::array::{
+    ArrayBuilder, ArrayRef, StructArray, StructBuilder, UInt64Array, UInt64Builder,
+};
+use arrow::datatypes::{DataType, Field, Schema, SchemaRef};
 use datafusion::logical_plan::LogicalPlan;
 use datafusion::physical_plan::ExecutionPlan;
 use uuid::Uuid;
 
 use super::protobuf;
-use crate::utils::PartitionStats;
 
 pub mod from_proto;
 pub mod to_proto;
@@ -57,6 +59,7 @@ impl PartitionId {
 pub struct PartitionLocation {
     pub partition_id: PartitionId,
     pub executor_meta: ExecutorMeta,
+    pub partition_stats: PartitionStats,
 }
 
 /// Meta-data for an executor, used when fetching shuffle partitions from other executors
@@ -84,6 +87,99 @@ impl From<protobuf::ExecutorMetadata> for ExecutorMeta {
             host: meta.host,
             port: meta.port as u16,
         }
+    }
+}
+
+/// Summary of executed partition
+#[derive(Debug, Copy, Clone)]
+pub struct PartitionStats {
+    num_rows: u64,
+    num_batches: u64,
+    num_bytes: u64,
+}
+
+impl Default for PartitionStats {
+    fn default() -> Self {
+        Self {
+            num_rows: 0,
+            num_batches: 0,
+            num_bytes: 0,
+        }
+    }
+}
+
+impl PartitionStats {
+    pub fn new(num_rows: u64, num_batches: u64, num_bytes: u64) -> Self {
+        Self {
+            num_rows,
+            num_batches,
+            num_bytes,
+        }
+    }
+
+    pub fn arrow_struct_repr(self) -> Field {
+        Field::new(
+            "partition_stats",
+            DataType::Struct(self.arrow_struct_fields()),
+            false,
+        )
+    }
+    fn arrow_struct_fields(self) -> Vec<Field> {
+        vec![
+            Field::new("num_rows", DataType::UInt64, false),
+            Field::new("num_batches", DataType::UInt64, false),
+            Field::new("num_bytes", DataType::UInt64, false),
+            Field::new("null_count", DataType::UInt64, false),
+        ]
+    }
+
+    pub fn to_arrow_arrayref(&self) -> Arc<StructArray> {
+        let mut field_builders = Vec::new();
+
+        let mut num_rows_builder = UInt64Builder::new(1);
+        num_rows_builder.append_value(self.num_rows).unwrap();
+        field_builders.push(Box::new(num_rows_builder) as Box<dyn ArrayBuilder>);
+
+        let mut num_batches_builder = UInt64Builder::new(1);
+        num_batches_builder.append_value(self.num_batches).unwrap();
+        field_builders.push(Box::new(num_batches_builder) as Box<dyn ArrayBuilder>);
+
+        let mut num_bytes_builder = UInt64Builder::new(1);
+        num_bytes_builder.append_value(self.num_bytes).unwrap();
+        field_builders.push(Box::new(num_bytes_builder) as Box<dyn ArrayBuilder>);
+
+        let mut struct_builder = StructBuilder::new(self.arrow_struct_fields(), field_builders);
+        struct_builder.append(true).unwrap();
+        Arc::new(struct_builder.finish())
+    }
+
+    pub fn from_arrow_struct_array(struct_array: &StructArray) -> PartitionStats {
+        return PartitionStats {
+            num_rows: struct_array
+                .column_by_name("num_rows")
+                .expect("from_arrow_struct_array expected a field num_rows")
+                .as_any()
+                .downcast_ref::<UInt64Array>()
+                .expect("from_arrow_struct_array expected num_rows to be a UInt64Array")
+                .value(0)
+                .to_owned(),
+            num_batches: struct_array
+                .column_by_name("num_batches")
+                .expect("from_arrow_struct_array expected a field num_batches")
+                .as_any()
+                .downcast_ref::<UInt64Array>()
+                .expect("from_arrow_struct_array expected num_batches to be a UInt64Array")
+                .value(0)
+                .to_owned(),
+            num_bytes: struct_array
+                .column_by_name("num_bytes")
+                .expect("from_arrow_struct_array expected a field num_bytes")
+                .as_any()
+                .downcast_ref::<UInt64Array>()
+                .expect("from_arrow_struct_array expected num_bytes to be a UInt64Array")
+                .value(0)
+                .to_owned(),
+        };
     }
 }
 
@@ -143,5 +239,9 @@ impl ExecutePartitionResult {
 
     pub fn path(&self) -> &str {
         &self.path
+    }
+
+    pub fn statistics(&self) -> &PartitionStats {
+        &self.stats
     }
 }
