+++ rust/scheduler/src/planner.rs
@@ -38,6 +38,7 @@ use datafusion::physical_plan::merge::MergeExec;
 use datafusion::physical_plan::ExecutionPlan;
 use log::{debug, info};
 use std::time::Instant;
+use tokio::task::JoinHandle;
 
 type SendableExecutionPlan = Pin<Box<dyn Future<Output = Result<Arc<dyn ExecutionPlan>>> + Send>>;
 type PartialQueryStageResult = (Arc<dyn ExecutionPlan>, Vec<Arc<QueryStageExec>>);
@@ -290,7 +291,6 @@ async fn execute_query_stage(
     );
 
     let partition_count = plan.output_partitioning().partition_count();
-    let mut meta = Vec::with_capacity(partition_count);
 
     let num_chunks = partition_count / executors.len();
     let num_chunks = num_chunks.max(1);
@@ -305,18 +305,8 @@ async fn execute_query_stage(
         partition_chunks.len()
     );
 
-    // build metadata for partition locations
-    for i in 0..partition_chunks.len() {
-        let executor_meta = &executors[i % executors.len()];
-        for part in &partition_chunks[i] {
-            meta.push(PartitionLocation {
-                partition_id: PartitionId::new(job_id, stage_id, *part),
-                executor_meta: executor_meta.clone(),
-            });
-        }
-    }
-
-    let mut executions = Vec::with_capacity(partition_count);
+    let mut executions: Vec<JoinHandle<Result<Vec<PartitionLocation>>>> =
+        Vec::with_capacity(partition_count);
     for i in 0..partition_chunks.len() {
         let plan = plan.clone();
         let executor_meta = executors[i % executors.len()].clone();
@@ -325,9 +315,18 @@ async fn execute_query_stage(
         executions.push(tokio::spawn(async move {
             let mut client =
                 BallistaClient::try_new(&executor_meta.host, executor_meta.port).await?;
-            client
-                .execute_partition(job_id, stage_id, partition_ids, plan)
-                .await
+            let stats = client
+                .execute_partition(job_id.clone(), stage_id, partition_ids.clone(), plan)
+                .await?;
+            let mut meta: Vec<PartitionLocation> = Vec::with_capacity(partition_ids.len());
+            for part in &partition_ids {
+                meta.push(PartitionLocation {
+                    partition_id: PartitionId::new(&job_id, stage_id, *part),
+                    executor_meta: executor_meta.clone(),
+                    partition_stats: *stats[*part].statistics(),
+                });
+            }
+            Ok(meta)
         }));
     }
 
@@ -335,11 +334,13 @@ async fn execute_query_stage(
     let results = futures::future::join_all(executions).await;
 
     // check for errors
+    let mut meta = Vec::with_capacity(partition_count);
     for result in results {
         match result {
             Ok(partition_result) => {
                 let final_result = partition_result?;
                 debug!("Query stage partition result: {:?}", final_result);
+                meta.extend(final_result);
             }
             Err(e) => {
                 return Err(BallistaError::General(format!(
